intranet/requirements.txt
Django==1.11
django-auth-ldap==1.7.0
django-filter==2.0.0
django-widget-tweaks==1.4.2
Pillow==5.2.0
pyasn1==0.4.4
pyasn1-modules==0.2.2
python-ldap==3.1.0
pytz==2018.5
